Issue #29: Separated code into more granuar consern space
